+++ src/main/java/org/mum/wap/presentation/controller/StartEventServlet.java
@@ -20,7 +20,7 @@ public class StartEventServlet extends HttpServlet {
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        long eventId = Long.parseLong(request.getParameter("eventId"));
+        int eventId = Integer.parseInt(request.getParameter("eventId"));
 
         EventService eventService = new EventService();
         eventService.startEvent(eventId);
